macuin/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
reportlab
openpyxl
python-docx
